feat: bring up terser
