+++ package.json
@@ -31,6 +31,7 @@
   "license": "MIT",
   "dependencies": {
     "micromatch": "^4.0.2",
+    "terser": "^4.3.9",
     "uglify-js": "^3.6.0"
   },
   "devDependencies": {
